pyproject.toml
[project]
name = "start-social"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.4",
    "emailvalidator>=0.3",
    "fastapi[standard]>=0.116.1",
    "passlib[bcrypt]>=1.7.4",
    "phonenumbers>=9.0.11",
    "psycopg2-binary>=2.9.10",
    "pydantic-extra-types>=2.10.5",
    "pydantic-settings>=2.10.1",
    "pytest>=8.4.1",
    "python-jose[cryptography]>=3.5.0",
    "ruff>=0.12.9",
    "sqlalchemy>=2.0.43",
    "sqlmodel>=0.0.24",
]
